manual lockfile

+++ puzle/ulensdb.py
@@ -1,14 +1,12 @@
 import os
 import time
 import numpy as np
-from filelock import FileLock, SoftFileLock
 from pathlib import Path
 import logging
 
 from puzle.utils import return_data_dir
 
 logger = logging.getLogger(__name__)
-# logging.getLogger('filelock').setLevel(logging.WARNING)
 
 
 def identify_is_nersc():
@@ -46,33 +44,46 @@ def load_db_ids():
     return db_ids
 
 
-def remove_db_id():
+def remove_db_id(retry_time=5):
     ulensdb_file_path = '%s/ulensdb/ulensdb.txt' % return_data_dir()
     lock_path = ulensdb_file_path.replace('.txt', '.lock')
-    lock = SoftFileLock(lock_path)
 
     my_db_id = fetch_db_id()
     if my_db_id is None:
         logger.debug(f'{my_db_id}: Skipping remove_db for local process')
         return
 
-    logger.debug(f'{my_db_id}: Attempting delete from {ulensdb_file_path}')
-    with lock:
-        db_ids = load_db_ids()
-        logger.debug(f'{my_db_id}: db_ids loaded {db_ids}')
-        db_ids.remove(my_db_id)
+    successFlag = False
+    while True:
+        time.sleep(abs(np.random.normal(scale=.01 * retry_time)))
+        logger.debug(f'{my_db_id}: Attempting delete from {ulensdb_file_path}')
+        if not os.path.exists(lock_path):
+            Path(lock_path).touch()
 
-        with open(ulensdb_file_path, 'w') as f:
-            for db_id in list(db_ids):
-                f.write('%s\n' % db_id)
+            db_ids = load_db_ids()
+            logger.debug(f'{my_db_id}: db_ids loaded {db_ids}')
+            db_ids.remove(my_db_id)
 
-    logger.debug(f'{my_db_id}: Delete success')
+            with open(ulensdb_file_path, 'w') as f:
+                for db_id in list(db_ids):
+                    f.write('%s\n' % db_id)
+
+            if os.path.exists(lock_path):
+                os.remove(lock_path)
+
+            successFlag = True
+
+        if successFlag:
+            logger.debug(f'{my_db_id}: Delete success')
+            return
+        else:
+            logger.debug(f'{my_db_id}: Delete fail, retry in {retry_time} seconds')
+            time.sleep(retry_time)
 
 
 def insert_db_id(num_ids=50, retry_time=5):
     ulensdb_file_path = '%s/ulensdb/ulensdb.txt' % return_data_dir()
     lock_path = ulensdb_file_path.replace('.txt', '.lock')
-    lock = SoftFileLock(lock_path)
 
     my_db_id = fetch_db_id()
     if my_db_id is None:
@@ -83,7 +94,8 @@ def insert_db_id(num_ids=50, retry_time=5):
     while True:
         time.sleep(abs(np.random.normal(scale=.01 * retry_time)))
         logger.debug(f'{my_db_id}: Attempting insert to {ulensdb_file_path}')
-        with lock:
+        if not os.path.exists(lock_path):
+            Path(lock_path).touch()
             db_ids = load_db_ids()
             if len(db_ids) < num_ids:
                 db_ids.add(my_db_id)
@@ -93,6 +105,8 @@ def insert_db_id(num_ids=50, retry_time=5):
                         f.write('%s\n' % db_id)
 
                 successFlag = True
+            if os.path.exists(lock_path):
+                os.remove(lock_path)
 
         if successFlag:
             logger.debug(f'{my_db_id}: Insert success')
@@ -100,3 +114,4 @@ def insert_db_id(num_ids=50, retry_time=5):
         else:
             logger.debug(f'{my_db_id}: Insert fail, retry in {retry_time} seconds')
             time.sleep(retry_time)
+
